Added Delete request for SchedDosages + moved to own file
